+++ packages/sst/src/constructs/ApiGatewayV1Api.ts
@@ -12,10 +12,7 @@ import * as apigV1AccessLog from "./util/apiGatewayV1AccessLog.js";
 import { App } from "./App.js";
 import { Stack } from "./Stack.js";
 import { Bucket } from "./Bucket.js";
-import {
-  BindingResource,
-  FunctionBindingProps,
-} from "./util/functionBinding.js";
+import { BindingResource, BindingProps } from "./util/binding.js";
 import { Duration, toCdkDuration } from "./util/duration.js";
 import { getFunctionRef, SSTConstruct, isCDKConstruct } from "./Construct.js";
 import { DnsValidatedCertificate } from "./cdk/dns-validated-certificate.js";
@@ -800,7 +797,7 @@ export class ApiGatewayV1Api<
   }
 
   /** @internal */
-  public getFunctionBinding(): FunctionBindingProps {
+  public getBindings(): BindingProps {
     return {
       clientPackage: "api",
       variables: {
