+++ packages/sst/src/constructs/Queue.ts
@@ -9,10 +9,7 @@ import {
   FunctionInlineDefinition,
   FunctionDefinition,
 } from "./Function.js";
-import {
-  BindingResource,
-  FunctionBindingProps,
-} from "./util/functionBinding.js";
+import { BindingResource, BindingProps } from "./util/binding.js";
 import { toCdkDuration } from "./util/duration.js";
 import { Permissions } from "./util/permission.js";
 
@@ -310,7 +307,7 @@ export class Queue extends Construct implements SSTConstruct {
   }
 
   /** @internal */
-  public getFunctionBinding(): FunctionBindingProps {
+  public getBindings(): BindingProps {
     return {
       clientPackage: "queue",
       variables: {
