+++ packages/sst/src/constructs/SsrSite.ts
@@ -88,9 +88,9 @@ import { Duration, toCdkDuration } from "./util/duration.js";
 import { Permissions, attachPermissionsToRole } from "./util/permission.js";
 import {
   BindingResource,
-  FunctionBindingProps,
+  BindingProps,
   getParameterPath,
-} from "./util/functionBinding.js";
+} from "./util/binding.js";
 import { useProject } from "../project.js";
 import { VisibleError } from "../error.js";
 import { RetentionDays } from "aws-cdk-lib/aws-logs";
@@ -1538,7 +1538,7 @@ function handler(event) {
   >;
 
   /** @internal */
-  public getFunctionBinding(): FunctionBindingProps {
+  public getBindings(): BindingProps {
     const app = this.node.root as App;
     return {
       clientPackage: "site",
